requirements.txt
# ===================================
# CRICKAI - REQUIREMENTS
# ===================================
# NOTE: Python 3.10 or 3.11 recommended
# Python 3.13 may have compatibility issues

# Data Processing
pandas>=2.0.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.66.0

# Deep Learning - TensorFlow
tensorflow>=2.15.0

# Flask API Backend
flask>=3.0.0
flask-cors>=4.0.0

# Visualization (optional)
matplotlib>=3.8.0
